+++ fractal_server/app/runner/v2/runner.py
@@ -2,11 +2,11 @@ from concurrent.futures import ThreadPoolExecutor
 from copy import copy
 from copy import deepcopy
 
+from ....images import Filters
 from ....images import SingleImage
 from ....images.tools import _filter_image_list
 from ....images.tools import find_image_by_path
 from ....images.tools import match_filter
-from .filters import Filters
 from .models import Dataset
 from .models import DictStrAny
 from .models import WorkflowTask
