+++ fractal_server/app/schemas/v2/dumps.py
@@ -14,8 +14,8 @@ from pydantic import BaseModel
 from pydantic import Extra
 from pydantic import root_validator
 
-from fractal_server.app.runner.v2.filters import Filters
 from fractal_server.app.schemas.v1.dumps import TaskDump as TaskDumpV1
+from fractal_server.images import Filters
 from fractal_server.images import SingleImage
