+++ tests/v2/03_api/test_api_dataset_images.py
@@ -1,6 +1,6 @@
 from devtools import debug
 
-from fractal_server.app.runner.v2.filters import Filters
+from fractal_server.images import Filters
 from fractal_server.images import SingleImage
 from fractal_server.images.tools import match_filter
