+++ tests/v2/test_benchmark_helper_functions.py
@@ -4,9 +4,9 @@ from time import perf_counter
 import pytest
 from devtools import debug
 
-from fractal_server.app.runner.v2.filters import Filters
 from fractal_server.app.runner.v2.runner_functions import deduplicate_list
 from fractal_server.app.runner.v2.runner_functions import InitArgsModel
+from fractal_server.images import Filters
 from fractal_server.images import SingleImage
 from fractal_server.images.tools import _filter_image_list
